Move EnvVar component to a separate file
